UserProfile code review changes
- Terms of Use re-signing is no longer forced, but the option is there
- Old UserProfile serialization method is fully deprecated (moved into test)
