+++ lib/jdomodels/src/main/java/org/sagebionetworks/repo/model/dbo/dao/SchemaSerializationUtils.java
@@ -2,7 +2,6 @@ package org.sagebionetworks.repo.model.dbo.dao;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
@@ -16,49 +15,6 @@ import org.sagebionetworks.schema.adapter.JSONObjectAdapterException;
 import org.sagebionetworks.schema.adapter.org.json.EntityFactory;
 
 public class SchemaSerializationUtils {
-	@SuppressWarnings("rawtypes")
-	public static byte[] mapDtoFieldsToAnnotations(Object dto, ObjectSchema schema) throws DatastoreException {
-		Map<String, ObjectSchema> schemaProperties = schema.getProperties();
-		NamedAnnotations properties = new NamedAnnotations();
-		Annotations a = properties.getPrimaryAnnotations();
-		for (String propertyName : schemaProperties.keySet()) {
-				try {
-					Field field = dto.getClass().getDeclaredField(propertyName);
-					field.setAccessible(true);
-					Map<String, List<String>> stringAnnots = a.getStringAnnotations();
-					Class fieldType = field.getType();
-					if (!(fieldType.equals(String.class) || fieldType.equals(AttachmentData.class))) {
-						// throw new RuntimeException("Unsupported field type "+fieldType);
-						continue; // Skip fields that are not supported
-					}
-					if (fieldType.equals(AttachmentData.class))
-					{
-						AttachmentData attachment = (AttachmentData)field.get(dto);
-						if (attachment != null)
-						{
-							stringAnnots.put(propertyName, Arrays.asList(new String[]{EntityFactory.createJSONStringForEntity(attachment)}));
-						}
-						else
-							stringAnnots.put(propertyName, Arrays.asList(new String[]{""}));
-					}
-					else //String
-						stringAnnots.put(propertyName, Arrays.asList(new String[]{(String)field.get(dto)}));
-				} catch (NoSuchFieldException e) {
-					// since the object is generated by the schema, this should never happen
-					throw new RuntimeException(e);
-				} catch (IllegalAccessException e) {
-					throw new RuntimeException(e);
-				} catch (JSONObjectAdapterException e) {
-					throw new RuntimeException(e);
-				}
-		}
-		try {
-			return JDOSecondaryPropertyUtils.compressAnnotations(properties);
-		} catch (IOException e) {
-			throw new DatastoreException(e);
-		}
-		
-	}
 	
 	@SuppressWarnings("rawtypes")
 	public static void mapAnnotationsToDtoFields(byte[] compressedProperties, Object dto, ObjectSchema schema) throws DatastoreException {
