+++ lib/jdomodels/src/test/java/org/sagebionetworks/repo/model/dbo/dao/UserProfileUtilsTest.java
@@ -4,16 +4,27 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.fail;
 
+import java.io.IOException;
+import java.lang.reflect.Field;
+import java.util.Arrays;
 import java.util.Date;
+import java.util.List;
+import java.util.Map;
 
 import org.junit.Test;
+import org.sagebionetworks.repo.model.Annotations;
+import org.sagebionetworks.repo.model.DatastoreException;
 import org.sagebionetworks.repo.model.Favorite;
+import org.sagebionetworks.repo.model.NamedAnnotations;
 import org.sagebionetworks.repo.model.UserProfile;
 import org.sagebionetworks.repo.model.attachment.AttachmentData;
 import org.sagebionetworks.repo.model.dbo.persistence.DBOFavorite;
 import org.sagebionetworks.repo.model.dbo.persistence.DBOUserProfile;
+import org.sagebionetworks.repo.model.jdo.JDOSecondaryPropertyUtils;
 import org.sagebionetworks.schema.ObjectSchema;
 import org.sagebionetworks.schema.adapter.JSONEntity;
+import org.sagebionetworks.schema.adapter.JSONObjectAdapterException;
+import org.sagebionetworks.schema.adapter.org.json.EntityFactory;
 import org.sagebionetworks.schema.adapter.org.json.JSONObjectAdapterImpl;
 
 public class UserProfileUtilsTest {
@@ -42,8 +53,7 @@ public class UserProfileUtilsTest {
 		
 		DBOUserProfile dbo = new DBOUserProfile();
 		UserProfileUtils.copyDtoToDbo(dto, dbo);
-		UserProfile dto2 = new UserProfile();
-		UserProfileUtils.copyDboToDto(dbo, dto2);
+		UserProfile dto2 = UserProfileUtils.convertDboToDto(dbo);
 		assertEquals(dto, dto2);
 	}
 	
@@ -79,13 +89,81 @@ public class UserProfileUtilsTest {
 		// Replace the blob with the older serialization method
 		String jsonString = (String) UserProfile.class.getField(JSONEntity.EFFECTIVE_SCHEMA).get(null);
 		ObjectSchema schema = new ObjectSchema(new JSONObjectAdapterImpl(jsonString));
-		dbo.setProperties(SchemaSerializationUtils.mapDtoFieldsToAnnotations(dto, schema));
+		dbo.setProperties(mapDtoFieldsToAnnotations(dto, schema));
 		
-		UserProfile dto2 = new UserProfile();
-		UserProfileUtils.copyDboToDto(dbo, dto2);
+		UserProfile dto2 = UserProfileUtils.convertDboToDto(dbo);
 		
 		dto.setAgreesToTermsOfUse(null); // Unsupported type
 		assertEquals(dto, dto2);
+	}
+
+	@Deprecated
+	@Test
+	public void testAnnotationsRoundtrip() throws Exception {
+		UserProfile dto = new UserProfile();
+		dto.setOwnerId("101");
+		dto.setFirstName("foo");
+		dto.setLastName("bar");
+		dto.setRStudioUrl("http://rstudio.com");
+		dto.setDisplayName("foo bar");
+		dto.setEtag("0");
+
+		String jsonString = (String) UserProfile.class.getField(JSONEntity.EFFECTIVE_SCHEMA).get(null);
+		ObjectSchema schema = new ObjectSchema(new JSONObjectAdapterImpl(jsonString));
+		byte[] na = mapDtoFieldsToAnnotations(dto, schema);
+		
+		UserProfile dto2 = new UserProfile();
+		SchemaSerializationUtils.mapAnnotationsToDtoFields(na, dto2, schema);
+
+		assertEquals(dto, dto2);
+	}
+
+	/**
+	 * Copied from SchemaSerializationUtils
+	 */
+	@Deprecated
+	@SuppressWarnings("rawtypes")
+	public static byte[] mapDtoFieldsToAnnotations(Object dto, ObjectSchema schema) throws DatastoreException {
+		Map<String, ObjectSchema> schemaProperties = schema.getProperties();
+		NamedAnnotations properties = new NamedAnnotations();
+		Annotations a = properties.getPrimaryAnnotations();
+		for (String propertyName : schemaProperties.keySet()) {
+				try {
+					Field field = dto.getClass().getDeclaredField(propertyName);
+					field.setAccessible(true);
+					Map<String, List<String>> stringAnnots = a.getStringAnnotations();
+					Class fieldType = field.getType();
+					if (!(fieldType.equals(String.class) || fieldType.equals(AttachmentData.class))) {
+						// throw new RuntimeException("Unsupported field type "+fieldType);
+						continue; // Skip fields that are not supported
+					}
+					if (fieldType.equals(AttachmentData.class))
+					{
+						AttachmentData attachment = (AttachmentData)field.get(dto);
+						if (attachment != null)
+						{
+							stringAnnots.put(propertyName, Arrays.asList(new String[]{EntityFactory.createJSONStringForEntity(attachment)}));
+						}
+						else
+							stringAnnots.put(propertyName, Arrays.asList(new String[]{""}));
+					}
+					else //String
+						stringAnnots.put(propertyName, Arrays.asList(new String[]{(String)field.get(dto)}));
+				} catch (NoSuchFieldException e) {
+					// since the object is generated by the schema, this should never happen
+					throw new RuntimeException(e);
+				} catch (IllegalAccessException e) {
+					throw new RuntimeException(e);
+				} catch (JSONObjectAdapterException e) {
+					throw new RuntimeException(e);
+				}
+		}
+		try {
+			return JDOSecondaryPropertyUtils.compressAnnotations(properties);
+		} catch (IOException e) {
+			throw new DatastoreException(e);
+		}
+		
 	}
 
 	@Test
@@ -99,8 +177,7 @@ public class UserProfileUtilsTest {
 		dto.setEtag("0");
 		DBOUserProfile dbo = new DBOUserProfile();
 		UserProfileUtils.copyDtoToDbo(dto, dbo);
-		UserProfile dto2 = new UserProfile();
-		UserProfileUtils.copyDboToDto(dbo, dto2);
+		UserProfile dto2 = UserProfileUtils.convertDboToDto(dbo);
 		assertEquals(dto, dto2);
 	}
